Issue: https://github.com/aelve/codesearch/issues/74
